♻️ refactor: 패키지 구조 변경으로 'accomplished' 모듈을 'challenge' 모듈로 이동
