+++ src/main/java/com/grow/member_service/challenge/accomplished/application/service/AccomplishedApplicationService.java
@@ -1,4 +1,4 @@
-package com.grow.member_service.accomplished.application.service;
+package com.grow.member_service.challenge.accomplished.application.service;
 
 import java.time.LocalDateTime;
 
@@ -7,11 +7,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.grow.member_service.accomplished.application.dto.AccomplishedResponse;
-import com.grow.member_service.accomplished.application.dto.CreateAccomplishedRequest;
-import com.grow.member_service.accomplished.application.model.AccomplishedPeriod;
-import com.grow.member_service.accomplished.domain.model.Accomplished;
-import com.grow.member_service.accomplished.domain.repository.AccomplishedRepository;
+import com.grow.member_service.challenge.accomplished.application.dto.AccomplishedResponse;
+import com.grow.member_service.challenge.accomplished.application.dto.CreateAccomplishedRequest;
+import com.grow.member_service.challenge.accomplished.application.model.AccomplishedPeriod;
+import com.grow.member_service.challenge.accomplished.domain.model.Accomplished;
+import com.grow.member_service.challenge.accomplished.domain.repository.AccomplishedRepository;
 import com.grow.member_service.common.exception.AccomplishedException;
 import com.grow.member_service.global.exception.ErrorCode;
