+++ src/main/java/com/grow/member_service/challenge/accomplished/infra/persistence/repository/AccomplishedRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.grow.member_service.accomplished.infra.persistence.repository;
+package com.grow.member_service.challenge.accomplished.infra.persistence.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -9,10 +9,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
-import com.grow.member_service.accomplished.domain.model.Accomplished;
-import com.grow.member_service.accomplished.domain.repository.AccomplishedRepository;
-import com.grow.member_service.accomplished.infra.persistence.entity.AccomplishedJpaEntity;
-import com.grow.member_service.accomplished.infra.persistence.mapper.AccomplishedMapper;
+import com.grow.member_service.challenge.accomplished.domain.model.Accomplished;
+import com.grow.member_service.challenge.accomplished.domain.repository.AccomplishedRepository;
+import com.grow.member_service.challenge.accomplished.infra.persistence.entity.AccomplishedJpaEntity;
+import com.grow.member_service.challenge.accomplished.infra.persistence.mapper.AccomplishedMapper;
 
 import lombok.RequiredArgsConstructor;
